requirements.txt
certifi==2019.11.28
cycler==0.10.0
imutils==0.5.3
kiwisolver==1.0.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
Pillow==7.0.0
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
tornado==6.0.3
wincertstore==0.2
opencv==3.4.2
matplotlib==3.0.2
